refactor: Refactor plugin services for modularity and clarity
Split install/uninstall and server metadata logic into new dedicated service classes. Move artwork and metadata handling to specialized static services. Update menu items and plugin initialization to use new services. Remove unused code from RommSyncService and update project file. Improves maintainability and separation of concerns.
